+++ src/fobi/contrib/apps/drf_integration/views.py
@@ -17,7 +17,7 @@ from ....constants import (
 from ....models import FormEntry
 
 from .base import (
-    # fire_form_callbacks,
+    fire_form_callbacks,
     run_form_handlers,
     submit_plugin_form_data,
 )
@@ -147,13 +147,13 @@ class FobiFormEntryViewSet(
         # Try to fetch only once.
         form_element_entries = form_entry.formelemententry_set.all()
 
-        # # Fire form valid before submit plugin data
-        # form = fire_form_callbacks(
-        #     form_entry=form_entry,
-        #     request=request,
-        #     form=form,
-        #     stage=CALLBACK_FORM_VALID_BEFORE_SUBMIT_PLUGIN_FORM_DATA
-        # )
+        # Fire form valid before submit plugin data
+        form = fire_form_callbacks(
+            form_entry=form_entry,
+            request=request,
+            serializer=serializer,
+            stage=CALLBACK_FORM_VALID_BEFORE_SUBMIT_PLUGIN_FORM_DATA
+        )
 
         # Fire plugin processors
         serializer = submit_plugin_form_data(
@@ -163,9 +163,12 @@ class FobiFormEntryViewSet(
         )
 
         # # Fire form valid callbacks
-        # form = fire_form_callbacks(form_entry=form_entry,
-        #                            request=request, form=form,
-        #                            stage=CALLBACK_FORM_VALID)
+        form = fire_form_callbacks(
+            form_entry=form_entry,
+            request=request,
+            serializer=serializer,
+            stage=CALLBACK_FORM_VALID
+        )
 
         # Run all handlers
         handler_responses, handler_errors = run_form_handlers(
@@ -187,9 +190,9 @@ class FobiFormEntryViewSet(
                 )
 
         # Fire post handler callbacks
-        # fire_form_callbacks(
-        #     form_entry=form_entry,
-        #     request=request,
-        #     form=form,
-        #     stage=CALLBACK_FORM_VALID_AFTER_FORM_HANDLERS
-        # )
+        fire_form_callbacks(
+            form_entry=form_entry,
+            request=request,
+            serializer=serializer,
+            stage=CALLBACK_FORM_VALID_AFTER_FORM_HANDLERS
+        )