+++ CHANGELOG.rst
@@ -15,6 +15,13 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.11.3
+------
+2017-05-10
+
+- Concept of integration callbacks introduced and implemented for the
+  ``drf_integration`` sub-package.
+
 0.11.2
 ------
 2017-05-09